+++ src/events.ts
@@ -2,7 +2,7 @@ import type { EventCallback } from '@tauri-apps/api/event';
 
 import { PaneEditorId, PaneId } from './atoms/types/PaneGroup';
 import { NotificationItemType } from './notifications/types';
-import { emit, listen } from './wrappers/tauri-wrapper';
+import { emit, listen } from './wrappers/listeners';
 
 interface RefStudioEvents {
   // Menu actions (Note that these should have undefined payload)
