+++ src/io/readEnv.ts
@@ -1,5 +1,3 @@
-// import { invoke } from '@tauri-apps/api';
-
 import { invoke } from '../wrappers/tauri-wrapper';
 
 export async function readEnv(key: string, fallback: string) {
