+++ src/vite-env.d.ts
@@ -1,1 +1,9 @@
 /// <reference types="vite/client" />
+
+interface ImportMetaEnv {
+  readonly VITE_IS_WEB?: string;
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv;
+}
