+++ src/wrappers/tauri-wrapper.ts
@@ -3,169 +3,32 @@ import * as tauriEvent from '@tauri-apps/api/event';
 import * as tauriFs from '@tauri-apps/api/fs';
 import * as tauriPath from '@tauri-apps/api/path';
 
-export const invoke: typeof tauriInvoke = (cmd, args) => {
-  console.log('invoke', cmd, args);
-  if (cmd === 'get_environment_variable') {
-    return '';
-  }
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-explicit-any
-  return null as any;
-};
+import * as stubEvent from './tauri-api-stubs/event';
+import * as stubFs from './tauri-api-stubs/fs';
+import { invokeStub } from './tauri-api-stubs/invokeStub';
+import * as stubPath from './tauri-api-stubs/path';
 
-// # @tauri-apps/api/path
-export const appConfigDir: typeof tauriPath.appConfigDir = () =>
-  Promise.resolve('/Users/refstudio/config/studio.ref.desktop');
-export const desktopDir: typeof tauriPath.desktopDir = () => Promise.resolve('/Users/refstudio/Desktop');
+// @tauri-apps/api
+export const invoke = import.meta.env.VITE_IS_WEB ? invokeStub : tauriInvoke;
 
-// XXX might need to resolve ".." and leading "/"
-export const join: typeof tauriPath.join = (...args) => Promise.resolve(args.join('/'));
+// @tauri-apps/api/path
+export const appConfigDir = import.meta.env.VITE_IS_WEB ? stubPath.appConfigDir : tauriPath.appConfigDir;
+export const desktopDir = import.meta.env.VITE_IS_WEB ? stubPath.desktopDir : tauriPath.desktopDir;
+export const join = import.meta.env.VITE_IS_WEB ? stubPath.join : tauriPath.join;
+export const step = import.meta.env.VITE_IS_WEB ? stubPath.sep : tauriPath.sep;
 
-export const sep: typeof tauriPath.sep = '/';
+// @tauri-apps/api/fs
+export const createDir = import.meta.env.VITE_IS_WEB ? stubFs.createDir : tauriFs.createDir;
+export const exists = import.meta.env.VITE_IS_WEB ? stubFs.exists : tauriFs.exists;
+export const readBinaryFile = import.meta.env.VITE_IS_WEB ? stubFs.readBinaryFile : tauriFs.readBinaryFile;
+export const readTextFile = import.meta.env.VITE_IS_WEB ? stubFs.readTextFile : tauriFs.readTextFile;
+export const readDir = import.meta.env.VITE_IS_WEB ? stubFs.readDir : tauriFs.readDir;
+export const removeDir = import.meta.env.VITE_IS_WEB ? stubFs.removeDir : tauriFs.removeDir;
+export const removeFile = import.meta.env.VITE_IS_WEB ? stubFs.removeFile : tauriFs.removeFile;
+export const writeTextFile = import.meta.env.VITE_IS_WEB ? stubFs.writeTextFile : tauriFs.writeTextFile;
+export const writeBinaryFile = import.meta.env.VITE_IS_WEB ? stubFs.writeBinaryFile : tauriFs.writeBinaryFile;
+export const renameFile = import.meta.env.VITE_IS_WEB ? stubFs.renameFile : tauriFs.renameFile;
 
-// # @tauri-apps/api/fs
-interface FakeFile {
-  type: 'file';
-  contents: string;
-}
-interface FakeDir {
-  type: 'dir';
-}
-type FakeNode = FakeFile | FakeDir;
-
-type NormalizedPath = string & { _brand: 'normalized' };
-// Remove some "warts" from paths, e.g. "//"
-function normalizePath(path: string): NormalizedPath {
-  let prefix = '';
-  if (path.startsWith('/')) {
-    prefix = '/';
-    path = path.slice(1);
-  }
-  return (prefix +
-    path
-      .split('/')
-      .filter((x) => !!x)
-      .join('/')) as NormalizedPath;
-}
-const fs = new Map<NormalizedPath, FakeNode>();
-
-export const createDir: typeof tauriFs.createDir = (dir, options) => {
-  if (!options?.recursive) {
-    throw new Error('Only recursive createDir is implemented');
-  }
-  fs.set(normalizePath(dir), { type: 'dir' });
-  return Promise.resolve();
-};
-
-export const exists: typeof tauriFs.exists = (file) => Promise.resolve(fs.has(normalizePath(file)));
-
-export const readBinaryFile: typeof tauriFs.readBinaryFile = async (filePath) => {
-  const f = await readTextFile(filePath);
-  return new TextEncoder().encode(f);
-};
-
-export const readTextFile: typeof tauriFs.readTextFile = (filePath) => {
-  const f = fs.get(normalizePath(filePath));
-  if (!f) {
-    throw new Error(`File does not exist: ${filePath}`);
-  }
-  if (f.type === 'dir') {
-    throw new Error(`Tried to read directory as file: ${filePath}`);
-  }
-  return Promise.resolve(f.contents);
-};
-
-export const readDir: typeof tauriFs.readDir = async (dir, options) => {
-  dir = normalizePath(dir);
-  const entries: tauriFs.FileEntry[] = [];
-  for (const [path, node] of fs.entries()) {
-    if (path.startsWith(dir) && path !== dir) {
-      entries.push({
-        path,
-        name: path.split('/').slice(-1)[0],
-        children: node.type === 'dir' ? await readDir(path, options) : undefined,
-      });
-    }
-  }
-  return entries;
-};
-
-export const removeDir: typeof tauriFs.removeDir = (dir, options) => {
-  if (!options?.recursive) {
-    throw new Error('Only recursive removeDir is available');
-  }
-  dir = normalizePath(dir);
-  for (const path of fs.keys()) {
-    if (path.startsWith(dir)) {
-      fs.delete(path);
-    }
-  }
-  return Promise.resolve();
-};
-
-export const removeFile: typeof tauriFs.removeFile = (filePath, options) => {
-  if (options) {
-    throw new Error('Not implemented.');
-  }
-  fs.delete(normalizePath(filePath));
-  return Promise.resolve();
-};
-
-export const writeTextFile: typeof tauriFs.writeTextFile = (filePathOrObj, contentsOrNone, ...args) => {
-  if (args.length) {
-    throw new Error('Not implemented.');
-  }
-  const filePath = typeof filePathOrObj === 'object' ? filePathOrObj.path : filePathOrObj;
-  const contents = typeof filePathOrObj === 'object' ? filePathOrObj.contents : (contentsOrNone as string);
-  fs.set(normalizePath(filePath), { type: 'file', contents });
-  return Promise.resolve();
-};
-
-export const writeBinaryFile: typeof tauriFs.writeBinaryFile = (filePathOrObj, contentsOrNone, ...args) => {
-  if (args.length) {
-    throw new Error('Not implemented.');
-  }
-  const filePath = typeof filePathOrObj === 'object' ? filePathOrObj.path : filePathOrObj;
-  const contents =
-    typeof filePathOrObj === 'object' ? filePathOrObj.contents : (contentsOrNone as tauriFs.BinaryFileContents);
-  fs.set(normalizePath(filePath), { type: 'file', contents: contents as string /* this is a lie */ });
-  return Promise.resolve();
-};
-
-export const renameFile: typeof tauriFs.renameFile = (oldPath, newPath, options) => {
-  if (options) {
-    throw new Error('Not implemented');
-  }
-  const normOldPath = normalizePath(oldPath);
-  const oldFile = fs.get(normOldPath);
-  if (!oldFile) {
-    throw new Error(`Tried to rename ${oldPath} which does not exist.`);
-  }
-  if (oldFile.type === 'dir') {
-    throw new Error('Renaming directores is not implemented');
-  }
-  fs.delete(normOldPath);
-  fs.set(normalizePath(newPath), oldFile);
-  return Promise.resolve();
-};
-
-(window as any).getFileSystem = () => fs;
-
-// # @tauri-apps/api/event
-const listeners = new Map<string, tauriEvent.EventCallback<unknown>>();
-
-export const emit: typeof tauriEvent.emit = (event, payload) => {
-  console.log('emit', event, payload);
-  const handler = listeners.get(event);
-  if (handler) {
-    handler({ event, payload, id: 0, windowLabel: 'refstudio' });
-  }
-  return Promise.resolve();
-};
-(window as any).emitEvent = emit;
-
-export const listen: typeof tauriEvent.listen = (event, callback) => {
-  listeners.set(event, callback as tauriEvent.EventCallback<unknown>);
-  return Promise.resolve(() => {
-    /* no op */
-  });
-};
+// @tauri-apps/api/event
+export const emit = import.meta.env.VITE_IS_WEB ? stubEvent.emit : tauriEvent.emit;
+export const listen = import.meta.env.VITE_IS_WEB ? stubEvent.listen : tauriEvent.listen;
